Handle storybook-addon-apollo-client storyshot failure
